+++ cmake/ImportCore.cmake
@@ -16,7 +16,7 @@ macro(import_core)
       download_project(
         PROJ scine-core
         GIT_REPOSITORY      https://github.com/qcscine/core.git
-        GIT_TAG             6.0.0
+        GIT_TAG             6.0.1
         QUIET
       )
       # Note: Options defined in the project calling this function override default
@@ -30,8 +30,8 @@ macro(import_core)
         )
       else()
         string(CONCAT error_msg
-          "Scine::Core was not found in your PATH and could not be established "
-          "through a download. Try specifying Scine_DIR or altering "
+          "Scine::Core was not found in your PATH and could not be downloaded. "
+          "Try specifying Scine_DIR or altering "
           "CMAKE_PREFIX_PATH to point to a candidate Scine installation base "
           "directory."
         )