+++ cmake/ImportSwoose.cmake
@@ -3,34 +3,36 @@
 # Copyright Department of Chemistry and Applied Biosciences, Reiher Group.
 # See LICENSE.txt for details.
 #
-macro(import_utils_os)
+macro(import_swoose)
   # If the target already exists, do nothing
-  if(NOT TARGET Scine::UtilsOS)
+  if(TARGET Scine::Swoose)
+    message(STATUS "Scine::Swoose present.")
+  else()
     # Try to find the package locally
-    find_package(ScineUtilsOS QUIET)
-    if(TARGET Scine::UtilsOS)
-      message(STATUS "Scine::UtilsOS found locally at ${ScineUtilsOS_DIR}")
+    find_package(ScineSwoose QUIET)
+    if(TARGET Scine::Swoose)
+      message(STATUS "Scine::Swoose found locally at ${ScineSwoose_DIR}")
     else()
       # Download it instead
       include(DownloadProject)
       download_project(
-        PROJ scine-utils-os
-        GIT_REPOSITORY https://github.com/qcscine/utilities.git
-        GIT_TAG        10.0.0
+        PROJ scine-swoose
+        GIT_REPOSITORY https://github.com/qcscine/swoose.git
+        GIT_TAG        2.1.0
         QUIET
       )
       # Note: Options defined in the project calling this function override default
       # option values specified in the imported project.
-      add_subdirectory(${scine-utils-os_SOURCE_DIR} ${scine-utils-os_BINARY_DIR})
+      add_subdirectory(${scine-swoose_SOURCE_DIR} ${scine-swoose_BINARY_DIR})
 
       # Final check if all went well
-      if(TARGET Scine::UtilsOS)
+      if(TARGET Scine::Swoose)
         message(STATUS
-          "Scine::UtilsOS was not found in your PATH, so it was downloaded."
+          "Scine::Swoose was not found in your PATH, so it was downloaded."
         )
       else()
         string(CONCAT error_msg
-          "Scine::UtilsOS was not found in your PATH and could not be downloaded. "
+          "Scine::Swoose was not found in your PATH and could not be downloaded. "
           "Try specifying Scine_DIR or altering "
           "CMAKE_PREFIX_PATH to point to a candidate Scine installation base "
           "directory."